+++ src/personas/preparePersona.ts
@@ -1,6 +1,7 @@
 import PipelineCollection from '../../promptbook-collection/index.json';
 import { createCollectionFromJson } from '../collection/constructors/createCollectionFromJson';
 import { IS_VERBOSE } from '../config';
+import { MissingToolsError } from '../errors/MissingToolsError';
 import { assertsExecutionSuccessful } from '../execution/assertsExecutionSuccessful';
 import { createPipelineExecutor } from '../execution/createPipelineExecutor/00-createPipelineExecutor';
 import type { PrepareAndScrapeOptions } from '../prepare/PrepareAndScrapeOptions';
@@ -20,6 +21,10 @@ export async function preparePersona(
     options: PrepareAndScrapeOptions,
 ): Promise<PersonaPreparedJson['modelRequirements']> {
     const { llmTools, isVerbose = IS_VERBOSE } = options;
+
+    if (llmTools === undefined) {
+        throw new MissingToolsError('LLM tools are required for preparing persona');
+    }
 
     // TODO: [🌼] In future use `ptbk make` and maked getPipelineCollection
     const collection = createCollectionFromJson(...(PipelineCollection as TODO_any as Array<PipelineJson>));