+++ CHANGELOG.md
@@ -529,6 +529,8 @@ Knowledge scrapers [🐝]
 -   Allow to import websites with knowledge
 -   Add new error `KnowledgeScrapeError`
 -   Add new error `getFilesystemToolsForNode`
+-   Remove `PipelineStringToJsonOptions` in favour of `PrepareAndScrapeOptions`
+-   Add `MissingToolsError`
 -   Add `FilesystemTools`
 -   Changed behavior of `titleToName` when passing URLs or file paths